+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Prevent the site's default category from being added to a post when using the
+  `with_terms()` method on a post factory if a category is being set.
+
 ## v1.5.5 - 2025-03-12
 
 No changes, just a re-release to fix a bad tag.